starry_core::shm

Struct ShmSegment

pub struct ShmSegment {
    pub id: ShmId,
    pub paddr: PhysAddr,
    pub size: usize,
    pub shmid_ds: Mutex<ShmidDs>,
    pub marked_for_deletion: AtomicBool,
}

Shared memory segment.

Fields
id: ShmId

Shared memory segment identifier.

paddr: PhysAddr

Physical address of the segment.

size: usize

Size of the segment in bytes.

shmid_ds: Mutex<ShmidDs>

Standard Linux shmid_ds structure (protected by mutex).

marked_for_deletion: AtomicBool

Whether this segment is marked for deletion.

Implementations

impl ShmSegment

pub fn new(id: ShmId, key: ShmKey, size: usize, mode: u16) -> AxResult<Self>

Creates a new shared memory segment.

pub fn inc_attach(&self)

Increments the attachment count for this segment.

pub fn dec_attach(&self)

Decrements the attachment count for this segment.

pub fn get_attach_count(&self) -> usize

Gets the current attachment count for this segment.

pub fn set_last_pid(&self, pid: i32)

Updates the last process ID that performed an operation.

pub fn check_permissions(&self, uid: u32, gid: u32, access: u16) -> bool

Checks if the given user has the required permissions for this segment.

pub fn validate(&self) -> bool

Validates that the segment is in a consistent state.

pub fn get_stat(&self) -> ShmidDs

Gets a copy of the shmid_ds structure for IPC_STAT.

pub fn set_perm(&self, uid: u32, gid: u32, mode: u32)

Updates permissions from user space (for IPC_SET).