starry_core

Module task

User task management.

Structs

ProcessData

Extended data for [Process].

TaskExt

Task extended data for the monolithic kernel.

ThreadData

Extended data for [Thread].

Functions

add_thread_to_table

Add the thread and possibly its process, process group and session to the corresponding tables.

get_process

Finds the process with the given PID.

get_process_group

Finds the process group with the given PGID.

get_session

Finds the session with the given SID.

get_thread

Finds the thread with the given TID.

new_user_task

Create a new user task.

processes

Lists all processes.

time_stat_from_kernel_to_user

Update the time statistics to reflect a switch from kernel mode to user mode.

time_stat_from_user_to_kernel

Update the time statistics to reflect a switch from user mode to kernel mode.

time_stat_output

Get the time statistics for the current task.